fix: display control labels in the default renderer as well for DashboardDataLayerSet
